integration_2023_w16b

See merge request oai/openairinterface5g!2069

* !1952 Fix problems detected by UBsan
* !1924 NR UL 256 QAM
* !1988 PUCCH2 for N = 64
* !2053 separating MAC and PHY procedures for TA
* !2058 fix interoperability with accelleran CU
* !2060 avoid duplicated function in nr_common for supported bandwidth
* !2064 CI AW2S config file fix
* !2065 bugfix for msg3 retransmission scheduling frame
* !2067 Modification of tagging develop images on OC

+++ openair2/RRC/NR/nr_rrc_proto.h
@@ -121,7 +121,7 @@ void nr_rrc_trigger(protocol_ctxt_t *ctxt, int CC_id, int frame, int subframe);
    \ *ptrsMcs Pointer to L_ptrs MCS related parameters
    \ *epre_Ratio Pointer to ep_ratio
    \ *reOffset Pointer to RE Offset Value */
-void rrc_config_dl_ptrs_params(NR_BWP_Downlink_t *bwp, int *ptrsNrb, int *ptrsMcs, int *epre_Ratio, int * reOffset);
+void rrc_config_dl_ptrs_params(NR_BWP_Downlink_t *bwp, long *ptrsNrb, long *ptrsMcs, long *epre_Ratio, long *reOffset);
 
 void nr_rrc_mac_remove_ue(rnti_t rntiP);
 void nr_rrc_mac_update_cellgroup(rnti_t rntiMaybeUEid, NR_CellGroupConfig_t *cgc);
